feat: Story time estimation + some graphs (#126)
* Semi-working graphs

* Fix Download PDF to be available to all users without login

* Fix bad auto-merge, manual fixes for story point estimation

+++ internal/database/db_user_story.go
@@ -9,8 +9,8 @@ type UserStory struct {
 	Title             string
 	Description       *string `gorm:"type:TEXT"`
 	Priority          Priority
-	BusinessValue     int
-	StoryPoints       uint
+	BusinessValue     uint
+	StoryPoints       float64
 	Realized          *bool
 	RejectionComment  *string `gorm:"type:TEXT"`
 	WorkflowStepID    *uint   // 1:1 (WorkflowStep:UserStory)
